[02/01/14] 1B024
- Pathfining is nearly done
- Added new mob with applied pathfinding

+++ src/com/mar21/au/entity/Entity.java
@@ -8,7 +8,7 @@ import com.mar21.au.level.Level;
 
 public class Entity {
 
-	protected double x, y;
+	protected int x, y;
 	protected Sprite sprite;
 	private boolean removed = false;
 	protected Level level;
@@ -52,11 +52,11 @@ public class Entity {
 		this.level = level;
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
